Reinput

+++ engine/ui/input/ui_input_types.h
@@ -52,8 +52,8 @@ struct UiInputEvent
     elysia::input::RawInputAxis axis = elysia::input::RawInputAxis::None;
     int mouse_x = 0;
     int mouse_y = 0;
-    int wheel_x = 0;
-    int wheel_y = 0;
+    float wheel_x = 0;
+    float wheel_y = 0;
     int composition_start = 0;
     int composition_length = 0;
     float axis_value = 0.0f;
